% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_cs.R
\name{cor_cs}
\alias{cor_cs}
\title{Calculate a more stringent similarity than correlation}
\usage{
cor_cs(xt)
}
\arguments{
\item{xt}{a matrix or data frame.}
}
\value{
Similarity dataframe
}
\description{
cor_cs computes the similarity between columns of a dataframe or matrix
}
