% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_cell_type.R
\name{pp_cell_type}
\alias{pp_cell_type}
\title{Pre-processing signature genes to return cell types present.}
\usage{
pp_cell_type(sig)
}
\arguments{
\item{exp}{List of cell types present in Signature table. Note: Gene names should be set as row names and duplicates are not allowed. Missing values are not allowed within the expression matrix. Check example- head(example_data): \code{\link{example_data}}.}
}
\value{
Signature Cell types list
}
\description{
List of cell types present in the Signature
}
