% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_cor.R
\name{pp_cor}
\alias{pp_cor}
\title{Pre-processing correlation matrix}
\usage{
pp_cor(cor_m, sig, cell_type)
}
\arguments{
\item{exp}{Dataframe of transcriptomic data (natural scale) containing genes as rows and samples as columns. Note: Gene names should be set as row names and duplicates are not allowed. Missing values are not allowed within the expression matrix. Check example- head(example_data): \code{\link{example_data}}.}
}
\value{
Correlation dataframe
}
\description{
Subsets the correlation dataset based on the genes that are common to Cell Signature and total correlation matrix.
}
